+++ src/main/java/ctrl/GroupV_Store/ProductController.java
@@ -5,20 +5,12 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.context.request.RequestContextHolder;
-import org.springframework.web.context.request.ServletRequestAttributes;
-
 import com.google.gson.Gson;
 
-import jakarta.servlet.http.HttpSession;
-import jakarta.websocket.Session;
 import models.Product;
 import models.ProductJdbcRepository;
-import models.Review;
 
 
 @RestController
@@ -51,21 +43,5 @@ public class ProductController {
 		}
 		Gson gson = new Gson();
 		return gson.toJson(products);
-	}
-	
-	@PostMapping("/reviews")
-	public void addReview(@RequestBody String data) {
-		Gson gson = new Gson();
-		Review review = gson.fromJson(data, Review.class);
-		
-		repository.addReview(review.getReviewer(), review.getStars(), review.getDetails(), review.getProductId());
-	}
-	
-	@GetMapping("/reviews")
-	public String getProductReviews(@RequestParam(value = "id") int productId) {
-		List<Review> reviews = repository.findReviews(productId);
-		
-		Gson gson = new Gson();
-		return gson.toJson(reviews);
 	}
 }
